+++ src/view/presenter/filmsBoard.js
@@ -6,11 +6,12 @@ import TitleNoFilmsView from "../films-no-title";
 import ButtonView from "../button";
 import FilmsListExtraView from "../films-list-extra";
 import FilmsBoardView from "../films-board";
-import {render, RENDER_POSITION, remove} from "../../util";
+
 import FilmView from "../film";
 import FilmPopupView from "../film-popup";
 import {TITLE_EXTRA} from "../../mock/extra-mock";
-const CARD_COUNT = 20;
+import {RenderPosition, render, remove} from "../../utils/render";
+
 const EXTRA_CARD_COUNT = 2;
 const SHOWN_CARDS_COUNT = 5;
 
@@ -27,7 +28,8 @@ export default class MovieList {
     this._button = new ButtonView();
     this._filmsListExtra = new FilmsListExtraView();
     this._handleButtonClickBind = this._handleButtonClick.bind(this);
-    this._shownCards = SHOWN_CARDS_COUNT;
+    this._shownCardsCount = SHOWN_CARDS_COUNT;
+    this._extraCardsCount = EXTRA_CARD_COUNT;
   }
   init(filmsData) {
     this._filmsData = filmsData.slice();
@@ -71,7 +73,7 @@ export default class MovieList {
       closeButton.removeEventListener(`click`, closeFilmPopup);
     };
     const openFilmPopup = () => {
-      render(this._popupContainer, FilmPopupViewComponent, RENDER_POSITION.BEFOREEND);
+      render(this._popupContainer, FilmPopupViewComponent, RenderPosition.BEFOREEND);
       closeButton.addEventListener(`click`, closeFilmPopup);
     };
 
@@ -92,7 +94,7 @@ export default class MovieList {
       }
     };
     document.addEventListener(`keydown`, onEscPress);
-    render(container, FilmCardComponent, RENDER_POSITION.BEFOREEND);
+    render(container, FilmCardComponent, RenderPosition.BEFOREEND);
   }
 
   _renderFilms(data, from, to, container) {
@@ -103,45 +105,45 @@ export default class MovieList {
   }
 
   _renderFilmsList() {
-    render(this._filmsBoard, this._filmsList, RENDER_POSITION.BEFOREEND);
+    render(this._filmsBoard, this._filmsList, RenderPosition.BEFOREEND);
   }
 
   _renderFilmsListContainer() {
-    render(this._filmsList, this._filmsListContainer, RENDER_POSITION.BEFOREEND);
+    render(this._filmsList, this._filmsListContainer, RenderPosition.BEFOREEND);
   }
 
   _renderFilmsTitle() {
-    render(this._filmsList, this._filmsTitle, RENDER_POSITION.AFTERBEGIN);
+    render(this._filmsList, this._filmsTitle, RenderPosition.AFTERBEGIN);
   }
 
   _renderFilmsNoTitle() {
-    render(this._filmsList, this._filmsNoTitle, RENDER_POSITION.AFTERBEGIN);
+    render(this._filmsList, this._filmsNoTitle, RenderPosition.AFTERBEGIN);
   }
   _handleButtonClick() {
 
-    this._renderFilms(this._filmsData, this._shownCards, this._shownCards + SHOWN_CARDS_COUNT, this._filmsListContainer);
-    this._shownCards += SHOWN_CARDS_COUNT;
-    if (this._shownCards >= CARD_COUNT) {
+    this._renderFilms(this._filmsData, this._shownCardsCount, this._shownCardsCount + SHOWN_CARDS_COUNT, this._filmsListContainer);
+    this._shownCardsCount += SHOWN_CARDS_COUNT;
+    if (this._shownCardsCount >= this._filmsData.length) {
       remove(this._filmsList, this._button);
       this._button.removeElement();
     }
   }
 
   _renderButton() {
-    render(this._filmsList, this._button, RENDER_POSITION.BEFOREEND);
+    render(this._filmsList, this._button, RenderPosition.BEFOREEND);
     this._button.setClickHandler(this._handleButtonClickBind);
   }
 
   _renderFilmsListExtra(data, title) {
     const filmsListExtraComponent = new FilmsListExtraView(title);
     const FilmsListContainerExtraComponent = new FilmsListContainerView();
-    render(this._filmsBoard, filmsListExtraComponent, RENDER_POSITION.BEFOREEND);
-    render(filmsListExtraComponent, FilmsListContainerExtraComponent, RENDER_POSITION.BEFOREEND);
+    render(this._filmsBoard, filmsListExtraComponent, RenderPosition.BEFOREEND);
+    render(filmsListExtraComponent, FilmsListContainerExtraComponent, RenderPosition.BEFOREEND);
     this._renderFilms(data, 0, EXTRA_CARD_COUNT, FilmsListContainerExtraComponent);
   }
 
   _renderFilmsBoard() {
-    render(this._movieListContainer, this._filmsBoard, RENDER_POSITION.BEFOREEND);
+    render(this._movieListContainer, this._filmsBoard, RenderPosition.BEFOREEND);
   }
 }
